Adding filter structure, navigation between issues
