+++ src/renderer/components/Validator/Report/convertToDataset.tsx
@@ -2,17 +2,28 @@ import {
     ITableData,
     ParsedValidationReport,
     DatasetJsonMetadata,
+    BasicFilter,
+    IUiValidationPage,
 } from 'interfaces/common';
+import Filter from 'js-array-filter';
 import renderRuleStatus from 'renderer/components/Validator/Report/Formatters/RuleStatus';
 import renderExecutability from 'renderer/components/Validator/Report/Formatters/Executable';
 import renderRow from 'renderer/components/Validator/Report/Formatters/Row';
 import renderVariables from 'renderer/components/Validator/Report/Formatters/Variables';
+import renderSummaryIssues from 'renderer/components/Validator/Report/Formatters/SummaryIssues';
 import columnDefs from 'renderer/components/Validator/Report/columnDefs';
 
 const convertToDataset = (
     data: ParsedValidationReport,
     type: 'Issue_Details' | 'Issue_Summary' | 'Rules_Report',
-    onOpenFile: (id: string) => void,
+    handlers: {
+        onOpenFile: (id: string) => void;
+        onFilterIssues: (
+            filter: BasicFilter,
+            reportTab: IUiValidationPage['currentReportTab'],
+        ) => void;
+    },
+    filter: BasicFilter | null = null,
 ): ITableData => {
     // Implement conversion logic here
     const metadata: DatasetJsonMetadata = {
@@ -24,6 +35,10 @@ const convertToDataset = (
         columns: [],
     };
 
+    let filtertedColumns: string[] = [];
+    if (filter) {
+        filtertedColumns = filter.conditions.map((c) => c.variable);
+    }
     let header: ITableData['header'] = [];
     // Get columns metadata
     if (type === 'Issue_Details') {
@@ -34,18 +49,18 @@ const convertToDataset = (
             dataType: col.type || 'string',
         }));
         header = columnDefs.details.map((col) => {
-            return col;
-        });
-        header = columnDefs.details.map((col) => {
-            const item = col;
+            let item = col;
+            if (filtertedColumns.includes(col.id)) {
+                item = { ...item, isFiltered: true };
+            }
             if (col.id === 'executability') {
                 item.cell = renderExecutability;
             }
             if (col.id === 'variables') {
-                item.cell = renderVariables(onOpenFile);
+                item.cell = renderVariables(handlers.onOpenFile);
             }
             if (col.id === 'row') {
-                item.cell = renderRow(onOpenFile);
+                item.cell = renderRow(handlers.onOpenFile);
             }
             return item;
         });
@@ -57,7 +72,14 @@ const convertToDataset = (
             dataType: col.type || 'string',
         }));
         header = columnDefs.summary.map((col) => {
-            return col;
+            let item = col;
+            if (filtertedColumns.includes(col.id)) {
+                item = { ...item, isFiltered: true };
+            }
+            if (col.id === 'issues') {
+                item.cell = renderSummaryIssues(handlers.onFilterIssues);
+            }
+            return item;
         });
     } else if (type === 'Rules_Report') {
         metadata.columns = columnDefs.rules.map((col) => ({
@@ -67,7 +89,10 @@ const convertToDataset = (
             dataType: col.type || 'string',
         }));
         header = columnDefs.rules.map((col) => {
-            const item = col;
+            let item = col;
+            if (filtertedColumns.includes(col.id)) {
+                item = { ...item, isFiltered: true };
+            }
             if (col.id === 'status') {
                 item.cell = renderRuleStatus;
             }
@@ -76,10 +101,23 @@ const convertToDataset = (
     }
 
     // Add row number
-    const updatedData = data[type].map((row, index) => ({
+    let updatedData = data[type].map((row, index) => ({
         ...row,
         '#': index + 1,
     }));
+    // Filter records;
+    if (filter) {
+        const filterInstance = new Filter(
+            'dataset-json1.1',
+            metadata.columns,
+            filter,
+        );
+        updatedData = updatedData.filter((row) => {
+            // We need to covert row to array;
+            const rowArray = metadata.columns.map((col) => row[col.name]);
+            return filterInstance.filterRow(rowArray);
+        });
+    }
 
     const result: ITableData = {
         header,
