+++ src/renderer/components/ViewerToolbar/index.tsx
@@ -102,7 +102,9 @@ const Header: React.FC = () => {
     }, [dispatch]);
 
     const handleFilterClick = useCallback(() => {
-        dispatch(openModal({ type: modals.FILTER, data: {} }));
+        dispatch(
+            openModal({ type: modals.FILTER, filterType: 'dataset', data: {} }),
+        );
     }, [dispatch]);
 
     const handleFilterReset = useCallback(() => {
